move getLatestMSBuild over to MSBuildTools
